backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pytesseract==0.3.10
opencv-python-headless==4.9.0.80
pdf2image==1.17.0
Pillow==10.3.0
pdfplumber==0.11.0
reportlab==4.1.0
aiosqlite==0.20.0
scikit-learn==1.4.2
numpy==1.26.4
httpx==0.27.0
pytest==8.1.1
pytest-asyncio==0.23.5
